Rename layout -> view-base
