+++ src/view-base/ColorView.js
