+++ src/view-base/HorizontalPanAndZoomView.js
