+++ src/view-base/ResizableSplitView.js
