+++ src/view-base/geometry.js
