+++ src/view-base/index.js
