+++ src/view-base/useCanvasInteraction.js
